prog: added move command(not working!)

+++ kernel/include/nop/type.h
@@ -18,6 +18,6 @@ typedef unsigned char      uint8_t;
 #endif
 
 // we need a signed 33-bit integer, but that doesn't exist so we have to do this...
-typedef int64_t off_t;
+typedef int32_t off_t;
 
 #endif
